requirements.txt
streamlit
python-docx
python-pptx
PyMuPDF
pandas
markdown
faiss-cpu       
sentence-transformers  
transformers